+++ CMakeLists.txt
@@ -27,12 +27,12 @@ add_executable(tidal
         src/app/DisplayManager.h
         src/app/models/DisplayData.cpp
         src/app/models/DisplayData.h
-        src/app/models/WeatherParser.cpp
-        src/app/models/WeatherParser.h
-        src/app/models/TideData.cpp
-        src/app/models/TideData.h
-        src/app/models/LunarData.cpp
-        src/app/models/LunarData.h
+        src/app/models/weather/WeatherParser.cpp
+        src/app/models/weather/WeatherParser.h
+        src/app/models/tide/TideData.cpp
+        src/app/models/tide/TideData.h
+        src/app/models/lunar/LunarData.cpp
+        src/app/models/lunar/LunarData.h
         src/app/TimeSeriesDataPoint.cpp
         src/app/TimeSeriesDataPoint.h
         src/app/config/AppConfig.cpp
@@ -49,18 +49,18 @@ add_executable(tidal
         src/app/config/ConfigFileNotFoundException.h
         src/app/FrameBufferOpenException.cpp
         src/app/FrameBufferOpenException.h
-        src/app/models/WeatherData.cpp
-        src/app/models/WeatherData.h
-        src/app/models/LunarPredictions.cpp
-        src/app/models/LunarPredictions.h
-        src/app/models/TidePredictions.cpp
-        src/app/models/TidePredictions.h
-        src/app/models/TideDataParser.cpp
-        src/app/models/TideDataParser.h
-        src/app/models/WeatherPredictions.cpp
-        src/app/models/WeatherPredictions.h
-        src/app/models/RawWeatherDatum.cpp
-        src/app/models/RawWeatherDatum.h)
+        src/app/models/weather/WeatherData.cpp
+        src/app/models/weather/WeatherData.h
+        src/app/models/lunar/LunarPredictions.cpp
+        src/app/models/lunar/LunarPredictions.h
+        src/app/models/tide/TidePredictions.cpp
+        src/app/models/tide/TidePredictions.h
+        src/app/models/tide/TideDataParser.cpp
+        src/app/models/tide/TideDataParser.h
+        src/app/models/weather/WeatherPredictions.cpp
+        src/app/models/weather/WeatherPredictions.h
+        src/app/models/weather/RawWeatherDatum.cpp
+        src/app/models/weather/RawWeatherDatum.h)
 
 include_directories( ${Boost_INCLUDE_DIR} /usr/local/include /usr/local/Cellar/openssl@1.1/1.1.1g/include/)
 target_include_directories(tidal PUBLIC "${PROJECT_BINARY_DIR}" /usr/local/include/lunar /usr/local/include/graphLib)
