+++ src/app/Application.h
@@ -8,7 +8,7 @@
 #include <lunar.h>
 #include <src/app/config/AppConfig.h>
 #include "DisplayManager.h"
-#include "src/app/models/LunarData.h"
+#include "src/app/models/lunar/LunarData.h"
 
 class Application {
 public:
