+++ src/app/models/weather/RawWeatherDatum.cpp
