+++ src/app/models/weather/WeatherData.cpp
