+++ src/app/models/weather/WeatherParser.cpp
@@ -55,3 +55,7 @@ std::map<std::string, std::vector<RawWeatherDatum>> WeatherParser::Parse(const s
 
     return predictions;
 }
+
+void WeatherParser::UnrollPredictions() {
+
+}
