+++ src/app/retrievers/LunarRetriever.h
@@ -5,7 +5,7 @@
 #ifndef TIDAL_LUNARRETRIEVER_H
 #define TIDAL_LUNARRETRIEVER_H
 
-#include <src/app/models/LunarData.h>
+#include <src/app/models/lunar/LunarData.h>
 #include <vector>
 #include <src/app/config/AppConfig.h>
