+++ tests/CMakeLists.txt
@@ -1,4 +1,4 @@
 
 # These tests can use the Catch2-provided main
-add_executable(tests RawWeatherDatumTests.cpp ../src/app/models/RawWeatherDatum.cpp)
+add_executable(tests RawWeatherDatumTests.cpp ../src/app/models/weather/RawWeatherDatum.cpp)
 target_link_libraries(tests PRIVATE Catch2::Catch2WithMain)
